ccl: move securityccl/fipsccl to security/fips
Informs #164337
Release note: None
