+++ pkg/BUILD.bazel
@@ -1063,7 +1063,6 @@ GO_TARGETS = [
     "//pkg/ccl/schemachangerccl/sctestbackupmrccl:sctestbackupmrccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl",
     "//pkg/ccl/schemachangerccl:schemachangerccl_test",
-    "//pkg/ccl/securityccl/fipsccl:fipsccl",
     "//pkg/ccl/serverccl/adminccl:adminccl_test",
     "//pkg/ccl/serverccl/diagnosticsccl:diagnosticsccl_test",
     "//pkg/ccl/serverccl/statusccl:statusccl_test",
@@ -1961,6 +1960,7 @@ GO_TARGETS = [
     "//pkg/security/distinguishedname:distinguishedname",
     "//pkg/security/dyncert:dyncert",
     "//pkg/security/dyncert:dyncert_test",
+    "//pkg/security/fips:fips",
     "//pkg/security/jwthelper:jwthelper",
     "//pkg/security/password:password",
     "//pkg/security/password:password_test",
