+++ pkg/security/fips/BUILD.bazel
@@ -1,13 +1,13 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "fipsccl",
+    name = "fips",
     srcs = [
         "fips_linux.go",
         "fips_nolinux.go",
         "sql.go",
     ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/securityccl/fipsccl",
+    importpath = "github.com/cockroachdb/cockroach/pkg/security/fips",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/ccl/utilccl",
